Ensure we respect the 10ms hard limit
Add documentation in CONTRIBUTING guide for ctx.Done() handling
Improve the graceful self-termination behavior of potentially
long-running signal goroutines by handling ctx.Done() timeouts.
This allows for the program to clean up, process defer'ed functions,
etc.

As a hard failsafe, implement a global watchdog in main() that hard
exits after 10.1ms if the program hasn't completed for any reason.

This is an important part of dashlight's design intent: the tool
should never, even under pathological system conditions, cause the
prompt to lag.

Finally, implement a make warning for signals that have not
implemented ctx.Done() checking. This *might* be acceptable depending
on what the signal is doing, but we should at least know when it is
being skipped.