Add Apple code signing and notarization to release process
Since we don't currently pay for GoReleaser Pro, manually sign and
notarize during release on a dedicated Macos runner. This adds
complexity to the actions workflow file, but is actually the
minimal set of steps needed.

NOTE: stapling isn't supported for raw binary/zipfiles, only APP,
DMG and PKG formats. This means that on the first execution after
download, users will make a network call to apple's notarization
servers to verify the binary. Future work may improve this with
an actual installer (e.g. to add the binary to path and insert
into prompt automatically), at which point installation would be
entirely offline.